Lots of patches and junk.

+++ sm_osx/RomView.swift
@@ -16,6 +16,14 @@ struct RomView: View {
     @State var log = ""
     @State var betterCamera = 0
     @State var drawDistance = 0
+    @State var doLog = false
+    @State var compSpeed: Speed = .fast
+    @State var extData = 0
+    #if arch(x86_64)
+    var disableCompilation = false
+    #else
+    var disableCompilation = true
+    #endif
     
     func shell(_ command: String) throws -> String {
         let task = Process()
@@ -55,7 +63,7 @@ struct RomView: View {
                         status = .instDependencies
                         
                         do {
-                            log = try shell("/usr/local/bin/brew install make mingw-w64 gcc gcc@9 sdl2 pkg-config glew glfw3 libusb audiofile coreutils")
+                            log = try shell("arch -x86_64 zsh && /usr/local/bin/brew install make mingw-w64 gcc gcc@9 sdl2 pkg-config glew glfw3 libusb audiofile coreutils && brew uninstall glew sdl2")
                         }
                         catch {
                             status = .rosetta
@@ -75,7 +83,7 @@ struct RomView: View {
                         status = .copyingFiles
                         
                         do {
-                            log.append(try shell("cd ~/Downloads && cp baserom.us.z64 sm64ex-coop && cd sm64ex-coop"))
+                            log.append(try shell("cd ~/Downloads && cp baserom.us.z64 sm64ex-coop"))
                             }
                         catch {
                             status = .error
@@ -86,7 +94,7 @@ struct RomView: View {
                         status = .compiling
                         
                         do {
-                            log.append(try shell("cd ~/Downloads/sm64ex-coop && gmake OSX_BUILD=1 TARGET_ARCH=x86_64-apple-darwin TARGET_BITS=64 BETTERCAMERA=\(betterCamera) NODRAWDISTANCE=\(drawDistance) -j6"))
+                            log.append(try shell("cd ~/Downloads/sm64ex-coop && gmake OSX_BUILD=1 TARGET_ARCH=x86_64-apple-darwin TARGET_BITS=64 BETTERCAMERA=\(betterCamera) NODRAWDISTANCE=\(drawDistance) EXTERNAL_DATA=\(extData) \(compSpeed.rawValue)"))
                         }
                         catch {
                             status = .error
@@ -97,7 +105,7 @@ struct RomView: View {
                         status = .finishingUp
                         
                         do {
-                            log.append(try shell("cd ~/Downloads && && rm -rf bin && gcp -r sm64ex-coop/build/us_pc/ bin"))
+                            log.append(try shell("cd ~/Downloads && rm -rf sm64ex-coop-build && gcp -r sm64ex-coop/build/us_pc/ sm64ex-coop-build"))
                         }
                         catch {
                             status = .error
@@ -115,17 +123,17 @@ struct RomView: View {
                         
                     }) {
                         Text("Start the Compiler")
-                    }.disabled(isCompiled)
+                    }.disabled(disableCompilation)
                 }
                 
-                else if repo == .sm64ex {
+                else if repo == .sm64ex || repo == .render96ex {
                     
                     Button(action:{
                         
                         status = .instDependencies
                         
                         do {
-                            log.append(try shell("brew install make mingw-w64 gcc sdl2 pkg-config glew glfw3 libusb audiofile"))
+                            log.append(try shell("brew install make mingw-w64 gcc sdl2 pkg-config glew glfw3 libusb audiofile coreutils"))
                         }
                         catch {
                             status = .notRosetta
@@ -136,7 +144,7 @@ struct RomView: View {
                         status = .instRepo
                         
                         do {
-                            log.append(try shell("cd ~/Downloads && rm -rf sm64ex && git clone https://github.com/EmeraldLoc/sm64ex.git"))
+                            log.append(try shell("cd ~/Downloads && rm -rf sm64ex && git clone \(repo.rawValue)"))
                         }
                         catch {
                             status = .error
@@ -144,29 +152,70 @@ struct RomView: View {
                             return
                         }
                         
-                        if patch.contains(.omm) {
-                            status = .patching
-                            
-                            do {
-                                log.append(try shell("cd ~/Downloads && git clone https://github.com/PeachyPeachSM64/sm64pc-omm.git && cp sm64pc-omm/patch/omm.patch sm64ex && rm -rf sm64pc-omm && cd sm64ex && git apply --reject --ignore-whitespace 'omm.patch'"))
-                            }
-                            catch {
-                                status = .error
+                        if repo == .sm64ex {
+                            if patch.contains(.omm) {
+                                status = .patching
                                 
-                                return
+                                do {
+                                    log.append(try shell("cd ~/Downloads && git clone https://github.com/PeachyPeachSM64/sm64pc-omm.git && cp sm64pc-omm/patch/omm.patch sm64ex && rm -rf sm64pc-omm && cd sm64ex && git apply --reject --ignore-whitespace 'omm.patch'"))
+                                }
+                                catch {
+                                    status = .error
+                                    
+                                    return
+                                }
                             }
-                        }
                         
-                        if patch.contains(.highfps) {
-                            status = .patching
-                            
-                            do {
-                                log.append(try shell("cd ~/Downloads/sm64ex && cp enhancements/60fps_ex.patch 60fps_ex.patch && git apply --reject --ignore-whitespace '60fps_ex.patch'"))
-                            }
-                            catch {
-                                status = .error
+                            if patch.contains(.highfps) {
+                                status = .patching
                                 
-                                return
+                                do {
+                                    log.append(try shell("cd ~/Downloads/sm64ex && cp enhancements/60fps_ex.patch 60fps_ex.patch && git apply --reject --ignore-whitespace '60fps_ex.patch'"))
+                                }
+                                catch {
+                                    status = .error
+                                    
+                                    return
+                                }
+                            }
+                            
+                            if patch.contains(.timeTrials) {
+                                status = .patching
+                                
+                                do {
+                                    log.append(try shell("cd ~/Downloads/sm64ex && wget https://sm64pc.info/downloads/patches/time_trials.2.4.hotfix.patch && git apply --reject --ignore-whitespace 'time_trials.2.4.hotfix.patch'"))
+                                }
+                                catch {
+                                    status = .error
+                                    
+                                    return
+                                }
+                            }
+                            
+                            if patch.contains(.captainToadStars) {
+                                status = .patching
+                                
+                                do {
+                                    log.append(try shell("cd ~/Downloads/sm64ex && wget https://sm64pc.info/downloads/patches/captain_toad_stars.patch && git apply --reject --ignore-whitespace 'captain_toad_stars.patch'"))
+                                }
+                                catch {
+                                    status = .error
+                                    
+                                    return
+                                }
+                            }
+                            
+                            if patch.contains(.extMoveset) {
+                                status = .patching
+                                
+                                do {
+                                    log.append(try shell("cd ~/Downloads/sm64ex && wget https://sm64pc.info/downloads/patches/Extended.Moveset.v1.03b.sm64ex.patch && git apply --reject --ignore-whitespace 'Extended.Moveset.v1.03b.sm64ex.patch'"))
+                                }
+                                catch {
+                                    status = .error
+                                    
+                                    return
+                                }
                             }
                         }
                         
@@ -184,7 +233,7 @@ struct RomView: View {
                         status = .compiling
 
                         do {
-                            log.append(try shell("cd ~/Downloads/sm64ex && gmake OSX_BUILD=1 TARGET_ARCH=x86_64-apple-darwin TARGET_BITS=64 BETTERCAMERA=\(betterCamera) NODRAWDISTANCE=\(drawDistance) -j6"))
+                            log.append(try shell("cd ~/Downloads/sm64ex && gmake OSX_BUILD=1 BETTERCAMERA=\(betterCamera) EXTERNAL_DATA=\(extData) NODRAWDISTANCE=\(drawDistance) \(compSpeed.rawValue)"))
                         }
                         catch {
                             status = .error
@@ -195,7 +244,7 @@ struct RomView: View {
                         status = .finishingUp
                         
                         do {
-                            log.append(try shell("cd ~/Downloads && rm -rf bin && gcp -r sm64ex/build/us_pc/ bin"))
+                            log.append(try shell("cd ~/Downloads && rm -rf \(repo)-build && gcp -r sm64ex/build/us_pc/ \(repo)-build"))
                         }
                         catch {
                             status = .error
@@ -211,21 +260,38 @@ struct RomView: View {
                             
                             return
                         }
-                        
+
                         status = .finished
                         
                     }) {
                         Text("Start the Compiler")
-                    }
+                    }.disabled(!disableCompilation)
                 }
                 
-                Text("The app will freeze until compilation is finished. The commpilation may take 5-15 min")
+                Text("The app will freeze until compilation is finished. The commpilation may take 1-8 min")
                 
                 Text(status.rawValue)
                 
-                ScrollView {
-                    Text("\(log)")
-                        .frame(width: 550)
+                Toggle(isOn: $doLog) {
+                    Text("Log Data")
+                }
+                
+                Picker("Compilation Speed", selection: $compSpeed) {
+                    Text("Slow")
+                        .tag(Speed.slow)
+                    Text("Fast")
+                        .tag(Speed.fast)
+                    Text("Very Fast")
+                        .tag(Speed.veryFast)
+                    Text("Fastest")
+                        .tag(Speed.fastest)
+                }
+                
+                if doLog && status == .finished {
+                    ScrollView {
+                        TextEditor(text: $log)
+                            .disabled(true)
+                    }
                 }
                 
                 Spacer()
@@ -244,6 +310,13 @@ struct RomView: View {
                 }
                 else {
                     drawDistance = 0
+                }
+                
+                if patch.contains(.extData) {
+                    extData = 1
+                }
+                else {
+                    extData = 0
                 }
             }
         }
